Deprecate HierarchicalConfig and it's tracking of parent configs - #2365
A Config is a long lived object and is derived via subConfig a lot.
Keeping track of the parent it was derived, creates long-lived object chains which takes the GC longer to release them.
It can even lead to OOM if detekt get's embedded in an other application which reuses the top most Config object.
The property 'parentPath' of the Config interface can be used as a replacement for parent.key calls.